feature: Agregando logica del comprobante y logica adicional en el borrado de usuario y libros

+++ src/service/book.service.js
@@ -1,25 +1,43 @@
 const bookRepository = require('../repository/book.repository.js');
+const genericRepository = require('../repository/generic.repository');
+const checkoutRepository = require('../repository/checkout.repository.js');
 
 const createBook = async function(book){
-    const bookCreated = await bookRepository.createBook(book);
-    const respuesta = {
-        mensaje: 'Se creo el libro con exito',
-        url: 'http://localhost:3000/book/'+bookCreated.insertId
-    }
-    return respuesta;
-}
-const getBook = async function(){
-    const books = await bookRepository.getBook();
-    if(books.length != 0){
-        return books;
+    const bookAlreadyRegistred = await bookRepository.getBookByName(book.name);
+    if(bookAlreadyRegistred.length == 0){
+        await bookRepository.createBook(book);
+        const lastBookId = await genericRepository.getLastEntity('books');
+        const respuesta = {
+            mensaje: 'Se creo el usuario con exito',
+            url: 'http://localhost:3000/user/'+lastBookId[0].id
+        }
+        return respuesta;
     }
     else{
-        throw new Error('Libro no encontrado');
+        throw new Error('Ya existe un libro con ese nombre');
     }
+}
+const getBook = async function(){
+    let bookList = await bookRepository.getBook();
+    bookList.forEach(book => {
+        if(book.isForSale == 1){
+            book.isForSale = true;
+        }
+        else{
+            book.isForSale = false;
+        }
+    });
+    return bookList;
 }
 const getBookById = async function(id){
     const book = await bookRepository.getBookById(id);
     if(book.length != 0){
+        if(book.isForSale == 1){
+            book.isForSale = true;
+        }
+        else{
+            book.isForSale = false;
+        }
         return book;
     }
     else{
@@ -43,12 +61,23 @@ const modifyBook = async function(book){
 const deleteBook = async function(id){
     const book = await bookRepository.getBookById(id);
     if(book.length != 0){
-        await bookRepository.deleteBook(id);
-        const respuesta = {
-            mensaje: 'Se elimino el usuario con exito',
-            libro: book
+        const checkUserInCheckout = await checkoutRepository.getCheckoutByUserId(id);
+        if(checkUserInCheckout.length == 0){
+            await bookRepository.deleteBook(id);
+            const respuesta = {
+                mensaje: 'Se elimino el libro con exito',
+                libro: book
+            }
+            return respuesta;
         }
-        return respuesta;
+        else{
+            await bookRepository.deactivateUser(id);
+            const respuesta = {
+                mensaje: 'Se desactivo el libro para proximas compras con exito',
+                usuario: user
+            }
+            return respuesta;
+        }
     }
     else{
         throw new Error('Libro no encontrado');
