+++ src/service/user.service.js
@@ -1,13 +1,16 @@
 const userRepository = require('../repository/user.repository.js');
+const checkoutRepository = require('../repository/checkout.repository.js');
+const genericRepository = require('../repository/generic.repository');
 const User = require('../model/user.js');
 
 const createUser = async function(user){
     const userAlreadyRegistred = await userRepository.getUserByName(user.name);
     if(userAlreadyRegistred.length == 0){
-        const userCreated = await userRepository.createUser(user);
+        await userRepository.createUser(user);
+        const lastUserId = await genericRepository.getLastEntity('users');
         const respuesta = {
             mensaje: 'Se creo el usuario con exito',
-            url: 'http://localhost:3000/user/'+userCreated.insertId
+            url: 'http://localhost:3000/user/'+lastUserId[0].id
         }
         return respuesta;
     }
@@ -17,12 +20,27 @@ const createUser = async function(user){
 }
 
 const getUser = async function(){
-    return await userRepository.getUser();
+    let userList = await userRepository.getUser();
+    userList.forEach(user => {
+        if(user.isActive == 1){
+            user.isActive = true;
+        }
+        else{
+            user.isActive = false;
+        }
+    });
+    return userList;
 }
 
 const getUserById = async function(id){
     const user = await userRepository.getUserById(id);
     if(user.length != 0){
+        if(user.isActive == 1){
+            user.isActive = true;
+        }
+        else{
+            user.isActive = false;
+        }
         return user;
     }
     else{
@@ -46,12 +64,24 @@ const modifyUser = async function(user){
 const deleteUser = async function(id){
     const user = await userRepository.getUserById(id);
     if(user.length != 0){
-        await userRepository.deleteUser(id);
-        const respuesta = {
-            mensaje: 'Se elimino el usuario con exito',
-            usuario: user
+        const checkUserInCheckout = await checkoutRepository.getCheckoutByUserId(id);
+        if(checkUserInCheckout.length == 0){
+            await userRepository.deleteUser(id);
+            const respuesta = {
+                mensaje: 'Se elimino el usuario con exito',
+                usuario: user
+            }
+            return respuesta;
         }
-        return respuesta;
+        else{
+            await userRepository.deactivateUser(id);
+            const respuesta = {
+                mensaje: 'Se desactivo el usuario con exito',
+                usuario: user
+            }
+            return respuesta;
+        }
+        
     }
     else{
         throw new Error('Usuario no encontrado');
@@ -61,17 +91,23 @@ const login = async function(req){
     const user = new User(null,req.body.name,null,req.body.password);
     const userLogged =  await userRepository.login(user);
     if(userLogged.length != 0){
-        req.session.userLogged = userLogged[0].id;
+        if(userLogged[0].isActive){
+            req.session.userLogged = userLogged[0].id;
+        }
+        else{
+            throw new Error('Error de login: El usuario esta desactivado');
+        }
     }
     else{
         throw new Error('Error de login: Nombre o contraseña incorrectos');
     }
 }
+
 module.exports = {
     createUser,
     getUser,
     getUserById,
     modifyUser,
     deleteUser,
-    login
+    login,
 }
